Add files via upload
The problem statement is to predict the percentage of marks that a student is expected to score based upon the number of hours they studied. This is a simple Univariate Linear Regression Problem. With the importing of important python libraries followed by the reading of the text data file using pandas, I have implemented the Linear Regression Algorithm on the dataset. Have done some data exploration with data visualization to know the relationship between the independent and dependent variables. The evaluation has been done on our test dataset which results in good accuracy and less mean square error. Our model was trained efficiently and was able to predict the right values when feed with further unseen Data.